+++ src/reflections/model/mod.rs
@@ -1,7 +1,5 @@
 mod checkin;
 mod config;
-pub mod monthly;
 
 pub use checkin::*;
 pub use config::*;
-pub use monthly::*;