+++ src/reflections/repo/mod.rs
@@ -1,2 +1,1 @@
 mod checkin;
-pub mod monthly;